Fix mars camlib compilation warnings.
Resolves few signed-unsigned compare issues.

+++ camlibs/mars/mars.c
@@ -387,7 +387,7 @@ mars_routine (Info *info, GPPort *port, char param, int n)
 int
 histogram (unsigned char *data, unsigned int size, int *htable_r, int *htable_g, int *htable_b)
 {
-	int x;
+	unsigned int x;
 	/* Initializations */
 	for (x = 0; x < 0x100; x++) {
 		htable_r[x] = 0;
@@ -463,7 +463,7 @@ mars_white_balance (unsigned char *data, unsigned int size, float saturation,
 	}
 	GP_DEBUG("White balance (bright): r=%1d, g=%1d, b=%1d, fr=%1.3f, fg=%1.3f, fb=%1.3f\n", r, g, b, r_factor, g_factor, b_factor);
 	if (max_factor <= 2.5) {
-		for (x = 0; x < (size * 3); x += 3)
+		for (x = 0; x < ((int)size * 3); x += 3)
 		{
 			d = (data[x+0]<<8) * r_factor;
 			d >>=8;
@@ -507,7 +507,7 @@ mars_white_balance (unsigned char *data, unsigned int size, float saturation,
 	"White balance (dark): r=%1d, g=%1d, b=%1d, fr=%1.3f, fg=%1.3f, fb=%1.3f\n",
 				r, g, b, r_factor, g_factor, b_factor);
 
-	for (x = 0; x < (size * 3); x += 3)
+	for (x = 0; x < ((int)size * 3); x += 3)
 	{
 		d = (int) 0xff08-(((0xff-data[x+0])<<8) * r_factor);
 		d >>= 8;
@@ -526,7 +526,7 @@ mars_white_balance (unsigned char *data, unsigned int size, float saturation,
 	/* ------------------ COLOR ENHANCE ------------------ */
 
 	if(saturation > 0.0) {
-		for (x = 0; x < (size * 3); x += 3)
+		for (x = 0; x < ((int)size * 3); x += 3)
 		{
 			r = data[x+0]; g = data[x+1]; b = data[x+2];
 			d = (int) (r + g + b) /3.;
